requirements_riscos.txt
streamlit>=1.28.0
langchain>=0.1.0
langchain-google-genai>=1.0.0
langchain-core>=0.1.0
python-dotenv>=1.0.0
pandas>=2.0.0
plotly>=5.18.0
beautifulsoup4>=4.12.0
requests>=2.31.0
lxml>=4.9.0
pydantic>=2.0.0
